Expanding search for time axis for applying time zero to all modules
The creation of an alternative time axis data array label (delay relative time zero) is no longer bound to dimension Y. It tries to find the time dimension by itself and apply time zero to it. If the time axis is absent, it skips this action. For WESPE and ALS modules, it is helpful in case one forgets to switch time zero off for Bunch ID mode (this can freeze the app).